+++ src/commands/source-add.ts
@@ -11,8 +11,8 @@ import { parseRegistryRef } from "../registry/resolve";
 import type { InstalledStashEntry } from "../registry/types";
 import { detectStashRoot } from "../sources/providers/provider-utils";
 import { syncFromRef } from "../sources/providers/sync-from-ref";
-import { ensureWebsiteMirror, validateWebsiteInputUrl } from "../sources/providers/website";
 import type { AddResponse } from "../sources/types";
+import { ensureWebsiteMirror, validateWebsiteInputUrl } from "../sources/website-ingest";
 import { ensureWikiNameAvailable, validateWikiName } from "../wiki/wiki";
 import {
   auditInstallCandidate,